Create temporary ssh key in a random file
I was having problems with simultaneous connections to multiples servers
because this script generate the temporary ssh key a fixed file.

So that, I assign `ssh_file_key_tmp` with a random suffix string and
this variable is used to generate the temporal ssh key in a file with
that name. This files are removed at the end of the connections.