Make remote key removal more reliable
The deletion command previously failed if $ssh_pubkey contains a /, which it can do because the key contents are base64 encoded where / is in the encoding alphabet and also if an existing key is used (the output of ssh-add -L as the full path to the key file is in the key comment).

This switches to using pipe as an infrequently used character in this context.